Fix move states with edges. Move static functions to ECTransitionUtils.
