+++ code/richediting/src/main/java/org/fbme/ide/richediting/adapters/ecc/ECTransitionController.java
@@ -7,7 +7,6 @@ import org.fbme.scenes.controllers.diagram.ConnectionController;
 import org.jetbrains.annotations.Nullable;
 
 import java.awt.*;
-import java.awt.geom.AffineTransform;
 import java.awt.geom.QuadCurve2D;
 import java.util.function.BiFunction;
 import java.util.function.Function;
@@ -37,8 +36,8 @@ public class ECTransitionController implements ConnectionController<ECTransition
                 centre.translate(p.x - x, p.y - y);
                 Point sp = new Point(sourceBound.x + sourceBound.width / 2, sourceBound.y + sourceBound.height / 2);
                 Point tp = new Point(targetBound.x + targetBound.width / 2, targetBound.y + targetBound.height / 2);
-                Point sourcePoint = ECTransitionPathSynchronizer.crossBound(centre, sp, sourceBound);
-                Point targetPoint = ECTransitionPathSynchronizer.crossBound(centre, tp, targetBound);
+                Point sourcePoint = ECTransitionUtils.crossBound(centre, sp, sourceBound);
+                Point targetPoint = ECTransitionUtils.crossBound(centre, tp, targetBound);
                 return new ECTransitionPath(sourcePoint, centre, targetPoint);
             };
         }
@@ -47,12 +46,12 @@ public class ECTransitionController implements ConnectionController<ECTransition
 
     @Override
     public Function<Point, ECTransitionPath> getSourceTransformation(final ECTransitionPath path) {
-        return ns -> transform(path, ns, path.target);
+        return ns -> transformSource(path, ns);
     }
 
     @Override
     public Function<Point, ECTransitionPath> getTargetTransformation(final ECTransitionPath path) {
-        return nt -> transform(path, path.source, nt);
+        return nt -> transformTarget(path, nt);
     }
 
     @Override
@@ -68,29 +67,31 @@ public class ECTransitionController implements ConnectionController<ECTransition
         myHighlightColor = highlightColor;
     }
 
-    public ECTransitionPath transform(ECTransitionPath path, Point ns, Point nt) {
-        AffineTransform at = new AffineTransform();
-
-        int oldX = path.target.x - path.source.x;
-        int oldY = path.target.y - path.source.y;
-        int newX = nt.x - ns.x;
-        int newY = nt.y - ns.y;
-        double scale;
-        if (oldX == 0 && oldY == 0) {
-            scale = 1;
-        } else {
-            scale = Math.sqrt((newX * newX + newY * newY) / ((double) oldX * oldX + oldY * oldY));
-        }
-        at.translate(ns.x, ns.y);
-        at.rotate(-Math.atan2(newX * oldY - newY * oldX, newX * oldX + newY * oldY));
-        at.scale(scale, scale);
-        at.translate(-path.source.x, -path.source.y);
-
-        Point nc = new Point();
+    private ECTransitionPath transformSource(ECTransitionPath path, Point ns) {
+        Rectangle sourceBound = mySourceHandle.getBounds();
         Point centre = new Point(path.centre);
-        at.transform(centre, nc);
+        Point sp = new Point(sourceBound.x + sourceBound.width / 2, sourceBound.y + sourceBound.height / 2);
+        int dxSource = ns.x - sp.x ;
+        int dySource = ns.y - sp.y;
+        sourceBound.translate(dxSource, dySource);
+        Point newSource = ECTransitionUtils.crossBound(centre, ns, sourceBound);
+        return new ECTransitionPath(newSource, centre, path.target);
+    }
 
-        return new ECTransitionPath(ns, nc, nt);
+    private ECTransitionPath transformTarget(ECTransitionPath path, Point nt) {
+        Rectangle targetBound = myTargetHandle.getBounds();
+        Point centre = new Point(path.centre);
+        Point tp = new Point(targetBound.x + targetBound.width / 2, targetBound.y + targetBound.height / 2);
+        int dxTarget = nt.x - tp.x;
+        int dyTarget = nt.y - tp.y;
+        targetBound.translate(dxTarget, dyTarget);
+        Point newTarget = ECTransitionUtils.crossBound(centre, nt, targetBound);
+        return new ECTransitionPath(path.source, centre, newTarget);
+    }
+
+    public ECTransitionPath transform(ECTransitionPath path, Point ns, Point nt) {
+        ECTransitionPath transformSourcePath = transformSource(path, ns);
+        return transformTarget(transformSourcePath, nt);
     }
 
     @Override
